separate socket functionality
